refactor: 클래스 명 변경 등 리팩터링
- 기존 RoomescapeController의 네이밍이 페이지 매핑 목적과 어울리지 않아서 PageController로 변경.
- ReservationController에 @RequestMapping 추가로 중복 코드 제거
