+++ src/main/java/roomescape/controller/PageController.java
@@ -1,10 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class RoomescapeController {
+public class PageController {
     @GetMapping("/admin")
     String admin() {
         return "admin/index";
