+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
@@ -8,20 +8,24 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.Reservation;
+import roomescape.Reservations;
 
 @RestController
+@RequestMapping("/reservations")
 public class ReservationController {
 
     private final Reservations reservations = new Reservations();
     private final AtomicLong index = new AtomicLong(1);
 
-    @GetMapping("/reservations")
+    @GetMapping()
     ResponseEntity<List<Reservation>> read() {
         return ResponseEntity.ok(reservations.getReservations());
     }
 
-    @PostMapping("/reservations")
+    @PostMapping()
     ResponseEntity<Reservation> create(@RequestBody Reservation reservation) {
         Reservation newReservation = Reservation.toEntity(reservation, index.getAndIncrement());
         reservations.add(newReservation);
@@ -29,7 +33,7 @@ public class ReservationController {
         return ResponseEntity.ok().body(newReservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     ResponseEntity<Void> delete(@PathVariable Long id) {
         reservations.remove(id);
         return ResponseEntity.ok().build();
