+++ src/test/java/roomescape/PageControllerTest.java
@@ -7,7 +7,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class RoomescapeControllerTest {
+class PageControllerTest {
     @Test
     void 일단계() {
         RestAssured.given().log().all()
